# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Backend API
  version: 1.0.0
  description: |
    API contract for authentication and Redis-backed HTTP sessions.

    CSRF protection is enabled. Every unsafe request (POST, PUT, PATCH, DELETE)
    must echo the value of the `XSRF-TOKEN` cookie in the `X-XSRF-TOKEN` header,
    or the request is rejected with 403 before authentication is considered. The
    cookie is readable by JavaScript and is issued on any response from the API,
    including anonymous ones. Login and logout both replace it, so the value has
    to be read at request time rather than cached.

    The service serves two separate security chains. The session-authenticated
    application API under `/api` uses the cookie and CSRF rules above. The SCIM v2
    interface under `/scim/v2` is stateless and authenticated per request by an
    opaque connector bearer token, with no session, no CSRF and no cookie: a
    connector token buys nothing under `/api`, and a session buys nothing under
    `/scim/v2`. The operations under the Connectors tag are how an administrator
    mints and retires the credentials that chain accepts.

    **The SCIM namespace is behind a release gate that is now ON by default.** It was
    off while Users and Groups were being built, because the two are one release
    capability: a directory that could create Users but had no Groups could not
    express authority. Groups are complete, so the gate stops hiding a finished
...

...
servers:
...
paths:
...
  /api/admin/accounts:
    get:
      tags: [Administration]
      summary: List every registered account
      description: |
        Administrative review of who has access. Requires `ROLE_ADMIN`: an
        authenticated account with any other role is refused with 403, and the
        403 is decided by the filter chain before this operation runs.

        The response never contains a password hash. It is serialised from a
        projection that has no field for one, so the guarantee is structural
        rather than a matter of review discipline.
      operationId: listAccounts
      responses:
        "200":
          description: Every account, ordered by username.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AdminRoleRequired"
  /api/admin/accounts/{username}/disable:
    post:
      tags: [Administration]
      summary: Close an account to logins
      description: |
        Sets the account's `enabled` flag to false. A disabled account is refused
        at login with the same bare `401` as a wrong password.

        Leaves the lockout untouched: this is not a penalty, and it says nothing
        about the account's failure run. Idempotent — disabling an already
        disabled account writes nothing, still revokes its sessions, and answers
        200.

        Every session the account already holds is REVOKED. Its next request
        arrives unauthenticated and earns a `401`, rather than working on until
        the session expires. A refused disable (`409`) revokes nothing.
      operationId: disableUser
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          $ref: "#/components/responses/AdminAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AdminRoleRequired"
        "404":
          $ref: "#/components/responses/UnknownAccount"
        "409":
          $ref: "#/components/responses/UnsafeAccountChange"
  /api/admin/accounts/{username}/enable:
    post:
      tags: [Administration]
      summary: Reopen an account to logins
      description: |
        Sets `enabled` back to true. Does NOT lift a lockout the account is
        serving — that is a separate capability, `unlock`. Idempotent.

        Does not restore sessions `disable` revoked either: the account signs in
        again.
      operationId: enableUser
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          $ref: "#/components/responses/AdminAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AdminRoleRequired"
        "404":
          $ref: "#/components/responses/UnknownAccount"
  /api/admin/accounts/{username}/unlock:
    post:
      tags: [Administration]
      summary: End a lockout
      description: |
        Clears the failure run and the recorded `locked_at`, which is the only way
        a lockout ends: a lockout has no duration and never lifts on its own. Says
        nothing about `enabled`: a disabled account can be unlocked and stays
        disabled.

        Idempotent — an account serving no lockout is answered 200 with nothing
        written.
      operationId: unlockUser
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          $ref: "#/components/responses/AdminAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AdminRoleRequired"
        "404":
          $ref: "#/components/responses/UnknownAccount"
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: HTTP session created by a successful login.
    connectorBearer:
      type: http
      scheme: bearer
      description: |
        An opaque connector token issued by an administrator of this deployment,
        presented as `Authorization: Bearer <value>` on every SCIM resource request.
        Accepted in that header and nowhere else — never from a query string, a form
        body or a cookie. There is no authorization server and no grant flow.
  parameters:
...
    Username:
      name: username
      in: path
      required: true
      description: The identity to act on, by its `userName`.
      schema:
        type: string
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: The current value of the XSRF-TOKEN cookie, echoed back unchanged.
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: A valid authenticated session is required.
...
    AdminAccount:
      description: The account as it stands after the change.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AccountSummary"
    UnknownAccount:
      description: |
        No account has that username. The body is empty; the caller is already an
        administrator, so this reveals nothing the listing would not.
    UnsafeAccountChange:
      description: |
        The action would leave nobody able to reverse it — an account disabling
        itself, or the last enabled administrator. Not 403: the caller holds the
        right role, the action itself is refused.
...
    AdminRoleRequired:
      description: |
        The session is valid and was accepted; the account's role is what was
        refused. Distinct from 401, which means no usable session arrived at all —
        a client must not treat this as a signal to sign in again.
...
  schemas:
...
    AccountSummary:
      type: object
      required: [id, userName, admin, active, locked, hasPassword, createdAt]
      properties:
        id:
          type: string
          format: uuid
          description: |
            The identity's stable SCIM resource id, non-reassignable. The same id
            the SCIM interface serves this identity under, and the id the audit
            trail records — `userName` is a mutable attribute and never an
            identifier.
        userName:
          type: string
          description: |
            The login and display attribute. Mutable, and addressed by the
            operations below because it is what an administrator reads; anything
            that must keep pointing at the same identity uses `id`.
        admin:
          type: boolean
          description: |
            Whether this identity holds administrative authority. DERIVED from
            direct membership of the server-seeded Admin group, not stored: there
            is no role column, so this reports the same fact the login path
            derives and there is nothing for the two to disagree about.

            Membership changes take effect at the identity's next login. A session
            already issued carries the authority it was issued with.
        active:
          type: boolean
          description: |
            Administrative standing, and SCIM's own `active` attribute — the two
            were separate flags until the login identity and the SCIM User became
            one resource, and keeping both would have left one unreachable over
            the wire.

            Not decorative: an inactive identity is refused at login with the same
            bare `401` as a wrong password. Changed only through the disable and
            enable operations.
        locked:
          type: boolean
          description: |
            Whether a lockout is in force. Independent of `active` — a lockout is
            imposed automatically by a run of failed logins. It has no expiry and
            there is no field for one: the lockout stands until an administrator
            calls the unlock operation, so this flag is the whole lock state.
        hasPassword:
          type: boolean
          description: |
            Whether a credential is set at all. A SCIM-provisioned identity may
            exist with no password — it is a supported state, not an incomplete
            one — and such an identity can be read, can be put in a Group, and
            cannot log in. Reported because "no password was ever set" is
            otherwise indistinguishable from "the password is wrong", and only the
            first is fixed by a SCIM write.

            The value itself is never reported, here or anywhere: this schema has
            no field a hash could occupy.
        createdAt:
          type: string
          format: date-time
          description: |
            When the SCIM resource was created. No longer nullable: the timestamp
            lives on the resource row, which cannot exist without one.
...